Separates the bumps of the release and development versions
If a failure happens in the release process the re-run of the
packaging is done on the new release. Instead of repeating the two bumps
and messing with tags, the new approach works by stages. First a new tag
and a new release version is generated. Then, after successful release a
new development version must be manually triggered with gulp
prep-new-dev.