Fixed input bug
The input boxes in the "Initial values of parameters" section in the mm-configuration view loses focus if a timer is running.
To fix this I have ensured that the timer that checks if the coe is online only is running when the coe-simulation service is in use, e.g. in the simulation view.